+++ CHANGELOG.md
@@ -14,7 +14,8 @@ WARNING:  This release contains major breaking changes that will impact all user
 - Added support for checking multiple ELB instances at once by passing a comma separated list of ELB instance names in metrics-elb-full.rb
 - Added check-autoscaling-cpucredits.rb for checking T2 instances in autoscaling groups that are running low on CPU credits
 - Updated the fog and aws-sdk gems to the latest versions to improve performance, reduce 3rd party gem dependencies, and add support for newer AWS features.
-- Add metrics-ec2-filter to store node ids and count matching a given filter
+- Added metrics-ec2-filter to store node ids and count matching a given filter
+- Added check-ec2-filter to perform check on the number of instances matching a filter
 
 ### Fixed
 - Renamed autoscaling-instance-count-metrics.rb -> metrics-autoscaling-instance-count.rb to match our naming scheme